Update Issue 1
Added NotifyUserOfNewVersionOnly() method to UpdateChecker class. When the listener receives a call-back to IUpdateCheckListener.OnFoundNewerVersion(), they can now chose to call GetNewerVersion() to download and execute an installer, or they can call NotifyUserOfNewVersionOnly() to notify the user of the update and optionally open a website where they can download the update manually. This should be useful for notifying Mono users running on non-Windows platforms (i.e. Linux and Mac OS) of updates without dealing with Windows installer files.